Allows for "credentials.location" notation in property names
Creates an abstract CredentialsProperty class that is extended by
nested classes in each properties class.
