+++ spring-cloud-gcp-core/pom.xml
@@ -18,5 +18,9 @@
 			<groupId>com.google.cloud</groupId>
 			<artifactId>google-cloud-core</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.springframework</groupId>
+			<artifactId>spring-core</artifactId>
+		</dependency>
 	</dependencies>
 </project>
