+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-trace/src/main/java/org/springframework/cloud/gcp/trace/GcpTraceProperties.java
@@ -16,7 +16,7 @@
 package org.springframework.cloud.gcp.trace;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.core.io.Resource;
+import org.springframework.cloud.gcp.core.CredentialsProperty;
 
 /**
  * Stackdriver Trace Properties.
@@ -49,7 +49,7 @@ public class GcpTraceProperties {
 
 	private String projectId;
 
-	private Resource credentialsLocation;
+	private Credentials credentials;
 
 	/**
 	 * A utility method to determine x% of total memory based on Zipkin's AsyncReporter.
@@ -97,11 +97,14 @@ public class GcpTraceProperties {
 		this.projectId = projectId;
 	}
 
-	public Resource getCredentialsLocation() {
-		return this.credentialsLocation;
+	public Credentials getCredentials() {
+		return this.credentials;
 	}
 
-	public void setCredentialsLocation(Resource credentialsLocation) {
-		this.credentialsLocation = credentialsLocation;
+	public void setCredentials(Credentials credentials) {
+		this.credentials = credentials;
+	}
+
+	public class Credentials extends CredentialsProperty {
 	}
 }
